pyproject.toml
[tool.poetry]
name = "one_piece"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
bs4 = "^0.0.1"
PyPDF2 = "^2.10.4"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
